requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
GeoAlchemy2==0.17.1
greenlet==3.1.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
psycopg2==2.9.10
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==2.2.2
zipp==3.21.0
